Lots of refactoring!
Not sure that we've hit the mark yet ...
