+++ src/execution/execution.rs
@@ -10,14 +10,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 use std::ops;
-use crate::util::{Bottom,Top};
-use super::{EvmState,Section};
-use crate::evm::instruction::Instruction;
+use crate::bytecode::{Section};
+use crate::state::{EvmState};
+use crate::instruction::Instruction;
 use super::semantics::{execute,Outcome};
+use crate::util::{Bottom,Top};
+
 
 /// Simple alias since we're always dealing with concrete executions
 /// here.
-type Bytecode = super::Bytecode<Instruction>;
+type Bytecode = crate::bytecode::Bytecode<Instruction>;
 
 pub struct Execution<'a,T:EvmState+Clone+PartialEq> {
     /// The bytecode being executed by this execution.
