+++ src/il/compiler.rs
@@ -9,21 +9,10 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
 use crate::il::{BinOp, Region, Term};
-use crate::evm::{Assembly,Bytecode,Section};
-use crate::evm::instruction::{AssemblyInstruction,Instruction};
-use crate::evm::instruction::AbstractInstruction::*;
+use crate::bytecode::{Assembly,Bytecode,Section};
+use crate::instruction::{AssemblyInstruction,Instruction};
+use crate::instruction::AbstractInstruction::*;
 use crate::util::*;
