+++ tests/disassembler.rs
@@ -1,8 +1,8 @@
 use std::fs;
 use std::path::{PathBuf};
 use evmil::util::{FromHexString};
-use evmil::evm::legacy;
-use evmil::evm::{Assembly};
+use evmil::legacy;
+use evmil::bytecode::{Assembly};
 
 pub static TESTS_DIR: &str = "tests/files";
